requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-phone-verify==3.0.0
django-phonenumber-field==7.0.2
django-restframework==0.0.1
djangorestframework==3.14.0
idna==3.4
nexmo==2.5.2
phonenumbers==8.13.4
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
twilio==7.16.1
urllib3==1.26.14
wrapt==1.14.1
